// dist/es2015/renderer/webgl1/render/RenderSystem.d.ts
export declare const render: any;
export declare const sendAttributeData: (gl: WebGLRenderingContext, shaderIndex: number, program: WebGLProgram, geometryIndex: number, ProgramData: any, LocationData: any, GLSLSenderData: any, GeometryData: any, ArrayBufferData: any, GPUDetectData: any, VaoData: any) => void;